@@ -92,14 +92,14 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^19.2.14",
-    "@angular/common": "^19.2.14",
-    "@angular/compiler": "^19.2.14",
-    "@angular/core": "^19.2.14",
-    "@angular/forms": "^19.2.14",
-    "@angular/platform-browser": "^19.2.14",
-    "@angular/platform-browser-dynamic": "^19.2.14",
-    "@angular/router": "^19.2.14",
+    "@angular/animations": "^20.1.6",
+    "@angular/common": "^20.1.6",
+    "@angular/compiler": "^20.1.6",
+    "@angular/core": "^20.1.6",
+    "@angular/forms": "^20.1.6",
+    "@angular/platform-browser": "^20.1.6",
+    "@angular/platform-browser-dynamic": "^20.1.6",
+    "@angular/router": "^20.1.6",
     "core-js": "^3.15.2",
     "normalize.css": "8.0.1",
     "rxjs": "~7.4.0",
@@ -110,10 +110,10 @@
     "gitbook-plugin-responsive-iframes": "1.0.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "19.2.15",
-    "@angular/cli": "^19.2.15",
-    "@angular/compiler-cli": "^19.2.14",
-    "@angular/language-service": "^19.2.14",
+    "@angular-devkit/build-angular": "20.1.5",
+    "@angular/cli": "^20.1.5",
+    "@angular/compiler-cli": "^20.1.6",
+    "@angular/language-service": "^20.1.6",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^16.11.7",
@@ -137,7 +137,7 @@
     "karma-firefox-launcher": "1.3.0",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^19.2.2",
+    "ng-packagr": "^20.1.0",
     "ngx-deploy-npm": "2.1.0",
     "node-sass": "^9.0.0",
     "np": "7.5.0",
@@ -148,7 +148,7 @@
     "semantic-release": "^17.1.1",
     "ts-node": "~10.4.0",
     "tslint": "~6.1.3",
-    "typescript": "~5.6.0",
+    "typescript": "~5.8.3",
     "validate-commit-msg": "2.12.1",
     "yargs": "15.4.0"
   },
